refactor dialog directories
